docs(README): switch 'owner' to 'owner_address' in filter examples
The `owner` example is confusing because people usually refer to owners
by address (SHA256 of public key) but `owner` in a filter refer to the
full key (same as the field on the TX). This adjusts the example to use
`owner_address` since that is typically what users will want.