chore(docker-compose): remove restart on-failure limit
Previously we limited the services to 5 restarts to prevent flapping.
Unfortunately, the restart counter is only reset when services are
manually cycled. This means that on-failure:5 not only prevents
flapping, but also means services will eventually stop restarting
entirely if they run into periodic issues. On balance we thinking that
it's worth tolerating occasional service flapping for improved
robustness.