fix(data-item-indexer): await when saving data items PE-6549
This was overlooked in the initial implementation. As long as the node
is keeping up with the queue, it's not a problem. However, if it gets
behind, the lack of await will cause promises to pile up in memory. With
the await in place we are able to handle that at the queue level by
dropping messages instead of leaking promises (though, we don't
currently have this implemented).