refactor(sqlite graphql): adjust cursors to align with ClickHouse adapter PE-6482
In SQLite we sort results using a placeholder value for data item IDs
when data items are not available. In the forthcoming ClickHouse adapter
we instead sort results using a flag indicating whether or not the row
is a data item (the order is the same with both approaches). Until this
commit, GraphQL cursors in both implementations reflected this
difference. This change adjusts the SQLite cursors to match the
ClickHouse cursors so that we can more easily combine results in the
future. This is accomplished by converting placeholder data item ID
values to/from boolean data item indicators in the cursor.