fix(arns): use ARNS_RESOLVER_PRIORITY_ORDER instead of ARNS_RESOLVER_TYPE
Now that we have the composite provider, we can make arns resolution a little more robust. Instead of requireing the `ARNS_RESOLVER_TYPE`, an operator can speicfy the resolution order they prefer. If a resolver, nor gateway urls are provided, only AO will be used. We default the order to use resolvers if available, then fallback to AO and gateways.