perf(graphql): add configurable tag selectivity sorting PE-7054
This change adds support for configuring tag join order via a
TAG_SELECTIVITY environment variable containing JSON mapping tag names
to selectivity weights. This is used when constructing SQLite tag join
SQL to determine the order in which to perform the joins (most selective
first). It also alters the table used to order results. If no tag is
present with a selectivity weight of >= 0 and there is an owner or
target address specified on the query the transactions or data item
tables are used for sorting rather than the tags table. This prevents a
lot of index scaning when low selectivity tags are combined with high a
high selectivity owner/target address filter.