feat(docker-compose): add ClickHouse service PE-6869
Added a ClickHouse service to docker-compose.yaml. I considered several
approaches for enabling/disabling it, but ultimately settled on using a
docker-compose profile so that we don't have to create our own custom
ClickHouse image (needed for ENV var based approach). Use '--profile
clickhouse' to enable it and start ClickHouse.