fix(arns): change from a `Set` of names to a `Map` for `ArNSNameCache`
This allows us to `get` the name if we have it, while persisting the debounce. This pattern will align better with our ANT cache and avoids overloading `has` and allows use to use additional data of the base ArNS record in the future - specficially when we start enforcing undername limits.