fix(debounce): await in KvDebounceStore on cache misses
This is a trade-off. Previously, on a cache miss - the request would return `undefined` and not await the debouncing of the cache. This means clients would have to retry the name to get it, after the debounce period has passed and the name has been refetched. Instead - it feels better to await that debounce/refetch and return the next result. Thi smeans the connection will be held open for the debounce interval + the resolution time, but should keep the cache hydrated for other requests and require less retries.