feat(arns): add undername enforcement to arns resolution
This change enforces undername limits from the ARIO contract for a name. By default, nodes will respect the priority provided by an ANT, and fall back to lexical sorting. A utility function provided by the ar.io/sdk is used to determine the index of a name. If a name requested is beyond the underanme limit, a 402 status code will be returned. Operators can choose to disable undername limit enforcements if they wish.