feat(data): use ario-peers as the final fallback when retrieving data
The ario-peers data retrieval option uses the AR.IO gateway address
registry to retrieve data from peer gateways in the network. This adds
it as the last fallback for on-demand retrieval (used for incoming GET
request for data). That means nodes in the network may end up serving
less trusted data as a last resort, but in the typical case for known
transaction IDs, that is unlikely to happen.