tests: add ability to tag tests and tag flaky tests PE-7696
This adds an 'isTestFiltered' function takes an array of tags and checks
them against the comma separated lists in TEST_ONLY_TAGS and
TEST_SKIP_TAGS. When TEST_ONLY_TAGS is used, only tests with tags in
that list will be run. When TEST_SKIP_TAGS is used, any test with tag in
that list will be skipped.

This change also tags tests that are currently flaky and adjusts the
GitHub workflow so that they will be skipped.