perf(graphql clickhouse): use ID sorted ClickHouse table for ID queries PE-7889
Without an ID sorted table ClickHouse scans the entire result set to
retrieve TXs with the given IDs (slow...). This commit adds an ID sorted
table and modifies the GraphQL query logic to use it when an ID filter
is included.

Note: The ID sorted table takes up a significant amount of space. We may
optimize this in the future by reducing the amount of data stored in
that table and treating it more like an index into the 'transactions'
table. ClickHouse compresses data very efficiently though, so for now
the storage hit seems worth the ease of implementation.