perf(graphql clickhouse): optimize single owner and recipient queries on ClickHouse PE-7810
Modifies the ClickHouse GraphQL SQL query generation to use tables
sorted by owner_address and target when there is a single owner or
recipient specified in the GraphQL quer. We will generalize this to
multiple owners and recipients as well in the future, but because of how
sorting works, that is a larger change that requires running multiple
queries.