perf(parquet-exporter): filter and sort tags using tables with height indexes PE-7928
This adjusts the ParquetExporter to use the 'height' columns on
'stable_transactions' and 'stable_data_items' for filtering and sorting
when exporting 'stable_transaction_tags' and 'stable_data_item_tags' so
as to take advantage of 'height' indexes on those tables. It also
switches the JOINs to CROSS JOINs to ensure SQLite chooses an efficient
join order when executing the queries.