fix(tx-chunks-data-source): throw (and 404) if the first chunk is missing PE-8060
This was discovered when attempting to sync an ArDrive drive. Prior to
this change, if retrieving data from chunks was attempted and chunks
were not found, the gateway would return 504. This happened because data
headers were sent and streaming started before attempting to retrieve
the first chunk. This change modifies the chunk source to attempt to
retrieve the first chunk prior to streaming and throw if it's missing.