feat(data): base expiration on data access times PE-8058
Adds a contiguous metadata cache that tracks access times and adjusts
the filesystem contiguous data cleanup process to use access times from
the cache when determining which files to delete.

Upcoming changes will expand the metadata cache in support of
preferential caching for ArNS data.

Also refactors the code in the following ways:
- Adds key prefix support to the Redis KV store.
- Consistently use prefixes for Redis cache keys.
- Adds optional support for overwriting values in KV attributes store on
  set.