pref(data): pass ranges through to getData PE-8103
getData implementations support returning data for regions (offset +
size). This adjusts the data handler HTTP range request support to use
the getData region functionality rather than streaming all the data for
the ID to extract the ranges.

Note: for data handlers that do not have efficient region
implementations (e.g. the chunk sources) this will be less efficient,
but range requests that use those sources are uncommon and they can be
optimized in the future.