fix(docker-compose): don't pass PARQUET_PATH to ClickHouse scripts
When running in a containerized environment, the path the scripts needs
is the one inside the container not outside. This conflicts with
specifying using the environment variable to specify the path for the
Docker volume mount. The solution is to not pass PARQUET_PATH into the
container and let the ClickHouse use the default path instead (it's
correct in that context).